Visual identity update (#270)
Updates the website for the new visual identity. 


* 663: Add test settings panel to frontend

* Remove unneeded code an styles for testing

* Changes to test styles

* Update custom block styles

* Update style experiments for fonts

* Update typography to latest draft

* Use filter to change Sendgrid sending email address

* Updated tests for Work Sans vs. Source Sans 3

* Revert "Use filter to change Sendgrid sending email address"

This reverts commit 32c7b470e193b884b34b70a2e181910852bfdafd.

* change secondary font to Work Sans and add letter spacing

* Improved petition columns

* Add additional parameters to form dataLayer event

* Add gp_user_id to form_submission event

* Update single.php template

* 671: Set sending address for email

P4 overrides the sending address for emails sent through Sendgrid. When we start using Planet4's connection to sendgrid with our own account, we need to override the sending address to be from our own domain.

* 672: Disable auto translate in WPML

We suspect auto translate causes a bug that reverts translated pages back to the original language. The bug cannot be replicated, so this is an attempt to fix it without knowing for sure it will work.

* Remove noheader-template for pages

P4 has a setting to hide the header now, the separate template is not needed anymore.

* WP CLI script to update page headers with the new pattern

* Update header template to newest version

* Update style

* Bugfix: Set correct image URL in header pattern

* Add a WP CLI command to clean up old pages

* Show more search results when adding pages to menu

* Style corrections for mater theme

* Code cleanup

* Generate styles

* Cleanup Javascript

* Remove unused SCSS files